+++ R/tidyverse.R
@@ -23,11 +23,12 @@
 #'
 #' * `use_tidy_issue_template()`: creates a standard tidyverse issue template.
 #'
-#' * `use_tidy_support`: creates support resources document for GitHub repo
+#' * `use_tidy_support()`: creates support resources document for GitHub repo
 #'    using SUPPORT.md.
 #'
-NULL
 #' @name tidyverse
+NULL
+
 
 #' @export
 #' @rdname tidyverse
@@ -122,7 +123,6 @@ use_tidy_eval <- function() {
 
 #' @export
 #' @rdname tidyverse
-#' @inheritParams use_template
 use_tidy_contributing <- function() {
   check_uses_travis()
 
@@ -139,7 +139,6 @@ use_tidy_contributing <- function() {
 
 #' @export
 #' @rdname tidyverse
-#' @inheritParams use_template
 use_tidy_issue_template <- function() {
   check_uses_github()
 
@@ -156,10 +155,8 @@ use_tidy_issue_template <- function() {
 use_tidy_support <- function() {
   check_uses_github()
 
-  name <- project_name()
-
   use_template(
     "SUPPORT.md",
     "SUPPORT.md")
-  data = list(package = name)
+  data = list(package = project_name())
 }