+++ NEWS.md
@@ -1,5 +1,8 @@
 # usethis 1.1.0.9000
 
+
+* `use_readme_rmd()` and `use_readme_md()` work, in a similar fashion, for projects that are and are not a package (@jennybc, #131, #135).
+
 * `use_readme_rmd()` once again creates a pre-commit git hook, to help keep `README.Rmd` and `README.md` in sync (@PeteHaitch @jennybc, #41).
 
 # usethis 1.1.0