+++ NEWS.md
@@ -1,3 +1,5 @@
+# usethis 1.2.0.9000
+
 # usethis 1.2.0
 
 ## New functions